Changed data fetching logic; OnAuthChange only fetches user; Data fetched by pages
