+++ src/components/LoadingSpinner.tsx
@@ -1,6 +1,6 @@
 export default function LoadingSpinner() {
   return (
-    <div className="spinner-container">
+    <div className="mx-auto w-fit justify-center pt-20">
       <div className="loading-spinner"></div>
     </div>
   );
