+++ src/pages/get-premium/index.tsx
@@ -59,7 +59,6 @@ export default function GetPremiumPage() {
   }
 
   if (isLoading) return <LoadingSpinner />;
-
   return (
     <div className="flex flex-col space-y-4">
       <div
