+++ src/pages/index.tsx
@@ -12,26 +12,145 @@ import Grid from "@/components/Grid";
 import UnderConstruction from "@/components/UnderConstruction";
 import IntroModal from "@/components/IntroModal";
 import { useEffect, useState } from "react";
+import useUserStore from "@/stores/user";
+import useTimerStore from "@/stores/timer";
+import {
+  getTasksInYearService,
+  getUnarchivedTasksService,
+} from "@/services/tasks";
+import { getSettingsService } from "@/services/settings";
+const { Howler } = require("howler");
+import { v4 as uuidv4 } from "uuid";
+import LoadingSpinner from "@/components/LoadingSpinner";
 
 export default function Home() {
   // Global states
-  const { tasks, taskFormType, isTaskEditMenuOpen } = useTaskStore();
-  const { tasksInTheYear } = useGridStore();
-  const { isSettingOpen } = useSettingStore();
+  const { taskFormType, isTaskEditMenuOpen, clearAllTasks, setTaskArray } =
+    useTaskStore();
+  const { setTasksInTheYear } = useGridStore();
+  const {
+    isSettingOpen,
+    setPomodoroTimerMinutes,
+    setShortBreakTimerMinutes,
+    setLongBreakTimerMinutes,
+    setNumberOfPomodoroSessionsInCycle,
+    setAlarmRingtone,
+    setAlarmVolume,
+    setWeekStart,
+  } = useSettingStore();
+  const { user } = useUserStore();
+  const { setTimerMinutes, setRemainingDurationInMilliseconds } =
+    useTimerStore();
+  const { addToast } = useToastStore();
 
-  // Local state for Intro modal
+  // Local states
   const [isIntroModalOpen, setIsIntroModalOpen] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
 
+  // useEffect to fetch Grid, Settings and Unarchived tasks data
   useEffect(() => {
-    if (tasks.length === 0 && tasksInTheYear.length === 0) {
-      setIsIntroModalOpen(true);
-    }
-  }, []);
+    const fetchData = async () => {
+      if (user) {
+        try {
+          setIsLoading(true);
+
+          // POST request: Retrieve user's tasks for the year
+          const tasksInTheYear = await getTasksInYearService(user, {
+            year: Number(new Date().getFullYear()),
+          });
+
+          if (tasksInTheYear) {
+            // Manipulate result: Change date to local time
+            let edited_tasksInTheYear = [];
+
+            for (let i = 0; i < tasksInTheYear.length; i++) {
+              const task = tasksInTheYear[i];
+
+              const edited = {
+                dateOfSession: new Date(task.dateOfSession)
+                  .toLocaleString()
+                  .split(",")[0],
+                completedNumOfMinutes: task.completedNumOfMinutes,
+                taskName: task.taskName,
+                category_name: task.category_name,
+                category_colour: task.category_colour,
+              };
+
+              edited_tasksInTheYear.push(edited);
+            }
+
+            // Set state
+            setTasksInTheYear(edited_tasksInTheYear);
+          }
+
+          // POST request: Retrieve user's settings
+          const settings = await getSettingsService(user);
+
+          if (settings) {
+            setPomodoroTimerMinutes(settings.pomodoro_minutes);
+            setShortBreakTimerMinutes(settings.short_break_minutes);
+            setLongBreakTimerMinutes(settings.long_break_minutes);
+            setNumberOfPomodoroSessionsInCycle(
+              settings.number_of_sessions_in_a_cycle
+            );
+            setAlarmRingtone(settings.alarm_ringtone);
+            setAlarmVolume(settings.alarm_volume);
+            setWeekStart(settings.week_start);
+            setTimerMinutes(settings.pomodoro_minutes);
+            setRemainingDurationInMilliseconds(
+              settings.pomodoro_minutes * 1000 * 60
+            );
+            Howler.volume(settings.alarm_volume);
+          }
+
+          // POST request: Retrieve user's unarchived tasks
+          const unarchivedTasks = await getUnarchivedTasksService(user);
+
+          if (unarchivedTasks) {
+            clearAllTasks();
+            setTaskArray(unarchivedTasks);
+          }
+
+          // Open IntroModal
+          if (unarchivedTasks?.length === 0 && tasksInTheYear?.length === 0) {
+            setIsIntroModalOpen(true);
+          }
+
+          setIsLoading(false);
+        } catch (error) {
+          // Add toast notification
+          addToast({
+            uniqueId: uuidv4(),
+            className: "bg-red-50 text-red-700",
+            content: "Something went wrong. Please try again! 😫",
+          });
+        }
+      } else {
+        setTasksInTheYear([]);
+
+        setPomodoroTimerMinutes(25);
+        setShortBreakTimerMinutes(5);
+        setLongBreakTimerMinutes(15);
+        setNumberOfPomodoroSessionsInCycle(4);
+        setAlarmRingtone("buzzer");
+        setAlarmVolume(0.5);
+        setTimerMinutes(25);
+        setRemainingDurationInMilliseconds(25 * 1000 * 60);
+
+        clearAllTasks();
+
+        setIsIntroModalOpen(true);
+      }
+    };
+
+    fetchData();
+  }, [user]);
 
   function toggleIntroModalFalse() {
     setIsIntroModalOpen(false);
   }
 
+  if (isLoading) return <LoadingSpinner />;
   return (
     <div>
       <Head>
@@ -39,26 +158,17 @@ export default function Home() {
         <meta name="description" content="Generated by create next app" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
       </Head>
-      {/************ Modals *************/}
-      {/* Intro modal */}
+
+      {/* Pop ups */}
       {isIntroModalOpen && (
         <IntroModal toggleIntroModalFalse={toggleIntroModalFalse} />
       )}
-
-      {/* Settings Form */}
       {isSettingOpen && <SettingsForm />}
-
-      {/* Task Form */}
       {taskFormType && <TaskForm />}
-
-      {/* Task Edit Menu */}
       {isTaskEditMenuOpen && <TaskEditMenu />}
 
-      {/************ Non-Modals *************/}
-      {/* Grid section */}
+      {/* Main Page */}
       <Grid />
-
-      {/* Timer and Tasks section */}
       <div className="flex flex-col space-y-4 lg:flex-row lg:space-x-4 lg:space-y-0">
         <TimerContainer />
         <TaskContainer />
